Ajustes no projeto

+++ src/main/java/com/lembretes/Lembretes/controllers/LembreteCotroller.java
@@ -1,5 +1,7 @@
 package com.lembretes.Lembretes.controllers;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,16 +11,16 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-
 import com.lembretes.Lembretes.Models.Entities.Lembretes;
 import com.lembretes.Lembretes.Models.Repositories.LembreteRepository;
+import com.lembretes.Lembretes.Models.Services.lembreteService;
 
 @Controller
 public class LembreteCotroller {
 
 	@Autowired
 	LembreteRepository lembreteRepository;
-
+	lembreteService lembreteService;
 	@GetMapping("/home")
 	public String home() {
 		return "index";
@@ -26,17 +28,15 @@ public class LembreteCotroller {
 
 	@GetMapping("/cadastro")
 	public String showCreateForm(Model model) {
-		model.addAttribute("lembrete", new Lembretes());
+		model.addAttribute("lembretes", new Lembretes());
 		return "cadastro";
 	}
 
 	@PostMapping(path = "/cadastro")
-	public /*@ResponseBody*/ String salvarLembrete(/*@Valid*/ Lembretes lembrete) {
+	public /* @ResponseBody */ String salvarLembrete(/* @Valid */ Lembretes lembrete) {
 		lembreteRepository.save(lembrete);
 		return "redirect:/home";
 	}
-
-	/* --------------------------------------------------------------- */
 
 	@RequestMapping("/ver")
 	public ModelAndView ver() {
@@ -46,43 +46,48 @@ public class LembreteCotroller {
 		return mv;
 	}
 
-	/*@GetMapping("/editar/{id}")
-	public ModelAndView editar(@PathVariable("id") int id){
-		ModelAndView mv = new ModelAndView("cadastro");
-		Lembretes lembreteFind = Lembretes.stream().filter(Lambretes ->id.equals)
- 		return mv;
-	}*/
+	@RequestMapping("/vizualizacao/{id}")
+	public ModelAndView vizualizacao( @PathVariable("id") Long id) {
+		ModelAndView mv = new ModelAndView("vizualizacao");
+		Optional<Lembretes> lembretes = lembreteRepository.findById(id);
+		mv.addObject("lembretes", lembretes);
+		return mv;
+	}
 
-	/*@DeleteMapping(path = "/excluir/{id}")
-    public void excluirProduto(@PathVariable int id){
-    	lembreteRepository.deleteById(id);
-    }*/
+	@GetMapping("/editar/{id}")
+	 public ModelAndView editar( @PathVariable("id") Long id) {
+		ModelAndView mv = new ModelAndView("editar");
+		Optional<Lembretes> lembretes = lembreteRepository.findById(id);
+		mv.addObject("lembretes", lembretes);
+		return mv;
+	}
+	/* --------------------------------------------------------------- */
 
-	/*@GetMapping("/ver")
-	public String pagever(){
-		return "verlembretes";
-	}*/
+	@GetMapping("/excluir/{id}")
+	 public String GetLembrete(@PathVariable("id") Long id) {
+        Optional<Lembretes> optionalLembrete = lembreteRepository.findById(id);
+		if (optionalLembrete.isPresent()){
+			lembreteRepository.deleteById(id);
+			return "redirect:/home";
+		}else {
+			return "redirect:/home";
+		}
+		//LembreteRepository.deleteById(id);
+        //return "redirect:/verlembretes";  // Redirecionar para a lista após deletar
+    }
 
-	
+	  @DeleteMapping("/excluir/{id}")
+	 public String deletarLembrete(@PathVariable("id") Long id) {
+        Optional<Lembretes> optionalLembrete = lembreteRepository.findById(id);
+		if (optionalLembrete.isPresent()){
+			lembreteRepository.deleteById(id);
+			return "redirect:/home";
+		}else {
+			return "redirect:/home";
+		}
+		//LembreteRepository.deleteById(id);
+        //return "redirect:/verlembretes";  // Redirecionar para a lista após deletar
+    }
 
-    /*@GetMapping("/ver")
-    public String listarLembretes(Model model) {
-        List<Lembretes> lembretes = lembreteRepository.findAll();
-        model.addAttribute("lembretes", lembretes);
-        return "verlembretes";
-    }*/
-	/*
-	 * @DeleteMapping(path = "/{id}")
-	 * public void excluirLembrete(@PathVariable int id) {
-	 * lembreteRepository.deleteById(id);
-	 * }
-	 */
-
-	/*
-	 * @GetMapping
-	 * public ArrayList<Lembretes> obterLembretes(){
-	 * return (ArrayList<Lembretes>) lembreteRepository.findAll();
-	 * }
-	 */
 
 }
